requirements.txt
streamlit
opencv-python
numpy
pandas
Pillow
easyocr